Refactor: final C daemon pre-commit cleanup and naming fixes.
